refactor: isMovablePoint 메서드 접근 제어자 변경 및 수정
