+++ src/main/java/chess/domain/piece/King.java
@@ -9,7 +9,8 @@ public final class King extends SingleMovePiece {
     }
 
     @Override
-    public boolean isMovablePoint(final Point departure, final Point destination) {
-        return departure.isAround(destination);
+    protected boolean isMovablePoint(final Point departure, final Point destination) {
+        return isNotSamePoint(departure, destination) &&
+                departure.isAround(destination);
     }
 }
