+++ src/main/java/chess/domain/piece/Queen.java
@@ -9,7 +9,8 @@ public final class Queen extends MultiMovePiece {
     }
 
     @Override
-    public boolean isMovableDirection(final Point departure, final Point destination) {
-        return departure.isStraight(destination) || departure.isDiagonalWithSlopeOfOne(destination);
+    protected boolean isMovablePoint(final Point departure, final Point destination) {
+        return isNotSamePoint(departure, destination) &&
+                (departure.isStraight(destination) || departure.isDiagonalWithSlopeOfOne(destination));
     }
 }
