+++ src/main/java/chess/domain/piece/SingleMovePiece.java
@@ -10,8 +10,6 @@ public abstract class SingleMovePiece extends Piece {
         super(type, team);
     }
 
-    public abstract boolean isMovablePoint(final Point departure, final Point destination);
-
     @Override
     public boolean isMovable(final Point departure, final Point destination, final Map<Point, Piece> board) {
         if (hasSameTeamPieceAtDestination(board.get(departure), board.get(destination))) {
