+++ src/test/java/chess/domain/piece/BishopTest.java
@@ -12,7 +12,7 @@ class BishopTest {
     @Test
     @DisplayName("오른쪽 대각선 위로 움직일 수 있다.")
     void isMovableDirection1() {
-        boolean result = bishop.isMovableDirection(
+        boolean result = bishop.isMovablePoint(
                 포인트("A1"),
                 포인트("B2"));
 
@@ -22,7 +22,7 @@ class BishopTest {
     @Test
     @DisplayName("오른쪽 대각선 아래로 움직일 수 있다.")
     void isMovableDirection2() {
-        boolean result = bishop.isMovableDirection(
+        boolean result = bishop.isMovablePoint(
                 포인트("C4"),
                 포인트("E2"));
 
@@ -32,7 +32,7 @@ class BishopTest {
     @Test
     @DisplayName("왼쪽 대각선 위로 움직일 수 있다.")
     void isMovableDirection3() {
-        boolean result = bishop.isMovableDirection(
+        boolean result = bishop.isMovablePoint(
                 포인트("H1"),
                 포인트("A8"));
 
@@ -42,7 +42,7 @@ class BishopTest {
     @Test
     @DisplayName("왼쪽 대각선 아래로 움직일 수 있다.")
     void isMovableDirection4() {
-        boolean result = bishop.isMovableDirection(
+        boolean result = bishop.isMovablePoint(
                 포인트("H8"),
                 포인트("B2"));
 
@@ -52,7 +52,7 @@ class BishopTest {
     @Test
     @DisplayName("오른쪽 일직선으로 움직일 수 없다.")
     void isNotMovableDirection1() {
-        boolean result = bishop.isMovableDirection(
+        boolean result = bishop.isMovablePoint(
                 포인트("A1"),
                 포인트("H1"));
 
@@ -62,7 +62,7 @@ class BishopTest {
     @Test
     @DisplayName("왼쪽 일직선으로 움직일 수 없다.")
     void isNotMovableDirection2() {
-        boolean result = bishop.isMovableDirection(
+        boolean result = bishop.isMovablePoint(
                 포인트("C4"),
                 포인트("A4"));
 
@@ -72,7 +72,7 @@ class BishopTest {
     @Test
     @DisplayName("위 일직선으로 움직일 수 없다.")
     void isNotMovableDirection3() {
-        boolean result = bishop.isMovableDirection(
+        boolean result = bishop.isMovablePoint(
                 포인트("A1"),
                 포인트("A8"));
 
@@ -82,9 +82,19 @@ class BishopTest {
     @Test
     @DisplayName("아래 일직선으로 움직일 수 없다.")
     void isNotMovableDirection4() {
-        boolean result = bishop.isMovableDirection(
+        boolean result = bishop.isMovablePoint(
                 포인트("H8"),
                 포인트("B8"));
+
+        assertThat(result).isFalse();
+    }
+
+    @Test
+    @DisplayName("같은 위치로 움직일 수 없다.")
+    void isNotMovableDirection5() {
+        boolean result = bishop.isMovablePoint(
+                포인트("H8"),
+                포인트("H8"));
 
         assertThat(result).isFalse();
     }
