+++ src/test/java/chess/domain/piece/QueenTest.java
@@ -13,7 +13,7 @@ class QueenTest {
     void isMovableDirection1() {
         Queen queen = new Queen(Team.WHITE);
 
-        boolean result = queen.isMovableDirection(
+        boolean result = queen.isMovablePoint(
                 Point.of('a', 1),
                 Point.of('a', 8)
         );
@@ -26,7 +26,7 @@ class QueenTest {
     void isMovableDirection2() {
         Queen queen = new Queen(Team.WHITE);
 
-        boolean result = queen.isMovableDirection(
+        boolean result = queen.isMovablePoint(
                 Point.of('a', 8),
                 Point.of('a', 1)
         );
@@ -39,7 +39,7 @@ class QueenTest {
     void isMovableDirection3() {
         Queen queen = new Queen(Team.WHITE);
 
-        boolean result = queen.isMovableDirection(
+        boolean result = queen.isMovablePoint(
                 Point.of('h', 1),
                 Point.of('a', 1)
         );
@@ -52,7 +52,7 @@ class QueenTest {
     void isMovableDirection4() {
         Queen queen = new Queen(Team.WHITE);
 
-        boolean result = queen.isMovableDirection(
+        boolean result = queen.isMovablePoint(
                 Point.of('a', 1),
                 Point.of('h', 1)
         );
@@ -64,7 +64,7 @@ class QueenTest {
     @DisplayName("오른쪽 대각선 위로 움직일 수 있다.")
     void isMovableDirection5() {
         Queen queen = new Queen(Team.WHITE);
-        boolean result = queen.isMovableDirection(
+        boolean result = queen.isMovablePoint(
                 Point.of('a', 1),
                 Point.of('b', 2)
         );
@@ -76,7 +76,7 @@ class QueenTest {
     @DisplayName("오른쪽 대각선 아래로 움직일 수 있다.")
     void isMovableDirection6() {
         Queen queen = new Queen(Team.WHITE);
-        boolean result = queen.isMovableDirection(
+        boolean result = queen.isMovablePoint(
                 Point.of('c', 4),
                 Point.of('e', 2)
         );
@@ -88,7 +88,7 @@ class QueenTest {
     @DisplayName("왼쪽 대각선 위로 움직일 수 있다.")
     void isMovableDirection7() {
         Queen queen = new Queen(Team.WHITE);
-        boolean result = queen.isMovableDirection(
+        boolean result = queen.isMovablePoint(
                 Point.of('h', 1),
                 Point.of('a', 8)
         );
@@ -100,7 +100,7 @@ class QueenTest {
     @DisplayName("왼쪽 대각선 아래로 움직일 수 있다.")
     void isMovableDirection8() {
         Queen queen = new Queen(Team.WHITE);
-        boolean result = queen.isMovableDirection(
+        boolean result = queen.isMovablePoint(
                 Point.of('h', 8),
                 Point.of('b', 2)
         );
